linguaphoto/requirements.txt
# requirements.txt

# Standard dependencies.
omegaconf
bson
pydantic

# AWS dependencies.
aioboto3
redis

# FastAPI dependencies.
aiohttp
aiosmtplib
fastapi
pyjwt
python-multipart

# Deployment dependencies.
uvicorn[standard]

# Types
types-aioboto3[dynamodb]

# AI dependencies: OpenAI and ElevenLabs.
elevenlabs
numpy
openai
Pillow
